Rename files and folders to use C++ conventions and for clarity
Add underscores to mathfunctions to result in math_functions

Create math_functions subdirectory in include directory to match source directory

Rename math_functions files to mysqrt files to differentiate from math_functions subdirectory name
