+++ Stage0/include/math_functions/mysqrt.h
@@ -1,0 +1,5 @@
+// define the custom namespace math_functions to group math functions and avoid naming conflicts
+namespace math_functions {
+// declare the custom square root function implemented in mysqrt.cpp
+double mysqrt(double x);
+}
