--- Stage0/include/mathfunctions.h
@@ -1,5 +1,0 @@
-// define the custom namespace mathfunctions to group math functions and avoid naming conflicts
-namespace mathfunctions {
-// declare the custom square root function implemented in mathfunctions.cpp
-double mysqrt(double x);
-}
